GitHub ScopingSearchHistoryMenu
